refactor: switch to static members on class
