+++ packages/server-wallet/e2e-test/e2e.test.ts
@@ -7,7 +7,7 @@ import {Channel} from '../src/models/channel';
 import {withSupportedState} from '../src/models/__test__/fixtures/channel';
 import {SigningWallet} from '../src/models/signing-wallet';
 import {stateVars} from '../src/wallet/__test__/fixtures/state-vars';
-import * as DBAdmin from '../src/db-admin/db-admin';
+import {DBAdmin} from '../src/db-admin/db-admin';
 
 import PayerClient from './payer/client';
 import {
