+++ packages/server-wallet/src/index.ts
@@ -3,8 +3,4 @@ export {Payload as Message} from '@statechannels/wallet-core';
 export * from './wallet';
 export {Outgoing} from './protocols/actions';
 
-// TODO: It would be more organized to export as DBAdmin
-// However this causes the api-extractor to fail
-// see https://github.com/microsoft/rushstack/issues/1029
-// export *  as DBAdmin from './db-admin/db-admin';
-export * from './db-admin/db-admin';
+export {DBAdmin} from './db-admin/db-admin';
