+++ packages/server-wallet/src/models/__test__/app-bytecode.test.ts
@@ -3,7 +3,7 @@ import {makeAddress} from '@statechannels/wallet-core';
 
 import {AppBytecode} from '../app-bytecode';
 import {testKnex as knex} from '../../../jest/knex-setup-teardown';
-import * as DBAdmin from '../../db-admin/db-admin';
+import {DBAdmin} from '../../db-admin/db-admin';
 
 const CHAIN_ID = '0x01';
 const APP_DEFINTION = makeAddress(constants.AddressZero);
