+++ packages/server-wallet/src/protocols/__test__/ledger-manager.test.ts
@@ -23,7 +23,7 @@ import {LedgerRequestType} from '../../models/ledger-request';
 import {getPayloadFor} from '../../__test__/test-helpers';
 import {ProposeLedgerUpdate} from '../actions';
 import {LedgerProposal} from '../../models/ledger-proposal';
-import {truncateDatabase} from '../../db-admin/db-admin';
+import {DBAdmin} from '../..';
 
 // TEST HELPERS
 // There are many test cases in this file. These helpers make the tests cases more readable.
@@ -60,11 +60,11 @@ beforeEach(async () => {
     logger: store.logger,
     timingMetrics: defaultTestConfig().metricsConfiguration.timingMetrics,
   });
-  await truncateDatabase(defaultTestConfig());
+  await DBAdmin.truncateDatabase(defaultTestConfig());
 });
 
 afterEach(async () => {
-  await truncateDatabase(defaultTestConfig());
+  await DBAdmin.truncateDatabase(defaultTestConfig());
 });
 
 describe('marking ledger requests as complete', () => {
