+++ packages/server-wallet/src/protocols/__test__/open-channel.test.ts
@@ -10,7 +10,7 @@ import {MockChainService} from '../../chain-service';
 import {createLogger} from '../../logger';
 import {DBOpenChannelObjective} from '../../models/objective';
 import {ChainServiceRequest, requestTimeout} from '../../models/chain-service-request';
-import {truncateDatabase} from '../../db-admin/db-admin';
+import {DBAdmin} from '../..';
 
 const logger = createLogger(defaultTestConfig());
 const timingMetrics = false;
@@ -25,11 +25,11 @@ beforeEach(async () => {
     defaultTestConfig().skipEvmValidation,
     '0'
   );
-  await truncateDatabase(defaultTestConfig());
+  await DBAdmin.truncateDatabase(defaultTestConfig());
 });
 
 afterEach(async () => {
-  await truncateDatabase(defaultTestConfig());
+  await DBAdmin.truncateDatabase(defaultTestConfig());
 });
 
 describe(`pre-fund-setup phase`, () => {
