+++ packages/server-wallet/src/wallet/__test__/challenge.test.ts
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
 import {defaultTestConfig, Wallet} from '..';
-import * as DBAdmin from '../../db-admin/db-admin';
+import {DBAdmin} from '../../db-admin/db-admin';
 import {seedAlicesSigningWallet} from '../../db/seeds/1_signing_wallet_seeds';
 import {AdjudicatorStatusModel} from '../../models/adjudicator-status';
 import {Channel} from '../../models/channel';
